\name{crossing}
\alias{crossing}
\title{G\'en\'eration d'un plan factoriel complet \`a partir des nombres de modalit\'es de...}
\usage{crossing(n, start=1)
}
\description{Génération d'un plan factoriel complet à partir des nombres de modalités de
\eqn{s} facteurs}
\value{matrice à 1 ligne par combinaison et 1 colonne par facteur}
\arguments{\item{n}{vecteur des nombres de modalités}
\item{start}{valeur du 1er chiffre utilisé pour les modalités des facteurs}
}
\examples{crossing( c(2,3,4), start=0 )}

