\name{ECmexico2012-package}
\alias{ECmexico2012-package}
\alias{ECmexico2012-package}
\title{Librairie R de l'ECOLE-CHERCHEURS MEXICO, ECULLY, 4-8 juin 2012}
\description{Librairie R de l'ECOLE-CHERCHEURS MEXICO, ECULLY, 4-8 juin 2012}
\details{\tabular{ll}{
Package: \tab ECmexico2012\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2012-05-24\cr
Depends: \tab R (>= 1.8.0), sensitivity, akima, car, evd, lhs, mgcv, rgl, triangle\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Collate: \tab 'ECmexico2012.R' 'TPLibModels.R' 'TPLibUtile.R' 'TPLibData.R'\cr
Encoding: \tab latin1\cr
}}
\docType{package}
\author{ec-mexico, maintainer: Hervé Monod}
\alias{ECmexico2012}
\alias{ECmexico2012-package}

