\name{fungus.factors}
\alias{fungus.factors}
\title{Facteurs d'entr\'ee de l'AS du mod\`ele "fungus"}
\description{Facteurs d'entrée du modèle "fungus" (croissance champignon selon l'humidité)}
\docType{data}
\value{data.frame à 5 lignes (facteurs) et 4 colonnes (spécifs)}
\note{Les valeurs dans fungus.factors correspondent à Alternaria Brassicae,
ravageur du colza (Magarey et al, 2005)}

