\name{inverses.basep}
\alias{inverses.basep}
\title{Calcul basique des inverses modulo p...}
\usage{inverses.basep(p)
}
\description{Calcul basique des inverses modulo \eqn{p}}
\value{vecteur des inverses}
\arguments{\item{p}{un nombre entier premier}
}
\examples{inverses.basep(5)
(inverses.basep(17) * (1:16)) \%\%17}

