\name{lhs2intervalle}
\alias{lhs2intervalle}
\title{Projection de valeurs tir\'ees entre 0 et 1 sur une autre plage de variation...}
\usage{lhs2intervalle(matrice, minAmax)
}
\description{Projection de valeurs tirées entre 0 et 1 sur une autre plage de variation}
\value{matrice}
\note{Utilisé dans lhs.plan.}
\arguments{\item{matrice}{matrice ou data.frame 0-1 à \eqn{N} lignes et \eqn{p} colonnes}
\item{minAmax}{vecteur des bornes de l'intervalle cible}
}
\examples{\dontrun{TODO}}

