\name{planor.kernelcheck.basep}
\alias{planor.kernelcheck.basep}
\title{V\'erification des confusions d'effets en cours de construction...}
\usage{planor.kernelcheck.basep(PhiStar, admissible, IneligibleSet, p)
}
\description{Vérification des confusions d'effets en cours de construction
d'un plan factoriel fractionnaire}
\value{vecteur logique}
\arguments{\item{PhiStar}{matrice clé en cours}
\item{admissible}{matrice codant les caractères a priori admissibles}
\item{IneligibleSet}{ensemble des caractères non éligibles}
\item{p}{un nombre entier premier}
}

