\name{representative.basep}
\alias{representative.basep}
\title{Fonction g\'en\'erant l'ensemble minimal de repr\'esentants des colonnes...}
\usage{representative.basep(mat, p)
}
\description{Fonction générant l'ensemble minimal de représentants des colonnes
d'une matrice, en base \eqn{p}}
\value{une matrice d'entiers modulo \eqn{p}}
\arguments{\item{mat}{une matrice d'entiers modulo \eqn{p}}
\item{p}{un nombre entier premier}
}

