\name{TP1.ICmorris}
\alias{TP1.ICmorris}
\title{Calcul d'intervalles de confiance pour la m\'ethode Morris...}
\usage{TP1.ICmorris(etude.morris)
}
\description{Calcul d'intervalles de confiance pour la méthode Morris}
\arguments{\item{etude.morris}{structure issue de la fonction morris du package sensitivity}
}

