\name{TP1corr}
\alias{TP1corr}
\title{Graphiques de corr\'elation entre facteurs X et sortie y de la FC...}
\usage{TP1corr(etude.morris, transfo=TRUE, binf, bsup)
}
\description{Graphiques de corrélation entre facteurs X et sortie y de la FC
issus d'une anova}
\arguments{\item{etude.morris}{structure issue de la fonction morris du package sensitivity}
\item{transfo}{si TRUE, recodage de la matrice X codée dans [0,1]}
\item{binf}{vecteur des bornes inférieures des gammes des facteurs}
\item{bsup}{vecteur des bornes supérieures des gammes des facteurs}
}

