\name{TP1histo}
\alias{TP1histo}
\title{Diagramme des fr\'equences des valeurs des facteurs \'echantillonn\'ees...}
\usage{TP1histo(etude.morris)
}
\description{Diagramme des fréquences des valeurs des facteurs échantillonnées
issus d'une anova}
\arguments{\item{etude.morris}{structure issue de la fonction morris du package sensitivity}
}

