\name{wwdm.climates}
\alias{wwdm.climates}
\alias{wwdm.climates}
\title{S\'eries climatiques sur 14 ann\'ees, utilis\'ees par le mod\`ele wwdm}
\description{Séries climatiques sur 14 années, utilisées par le modèle wwdm}
\docType{data}
\value{data.frame à N lignes (1 par jour) et 4 colonnes (ANNEE, RG, Tmin, Tmax)}

